minor: Condensed vcxproj.

+++ KeyView2/dbgprint.cpp
@@ -18,17 +18,17 @@ void dbgprint(const TCHAR *fmt, ...)
 	_sntprintf(buf, COUNT(buf)-3, TEXT("[%d] "), ++count); // prefix seq
 #endif
 	
-	int prefixlen = _tcslen(buf);
+	int prefixlen = (int)_tcslen(buf);
 	
 	va_list args;
 	va_start(args, fmt);
 #if _MSC_VER >= 1400 // VS2005+
 	_vsntprintf_s(buf+prefixlen, COUNT(buf)-3-prefixlen, _TRUNCATE, fmt, args);
-	prefixlen = _tcslen(buf);
+	prefixlen = (int)_tcslen(buf);
 	_tcsncpy_s(buf+prefixlen, 2, TEXT("\r\n"), _TRUNCATE); // add trailing \r\n
 #else
 	_vsntprintf(buf+prefixlen, COUNT(buf)-3-prefixlen, fmt, args);
-	prefixlen = _tcslen(buf);
+	prefixlen = (int)_tcslen(buf);
 	_tcsncpy(buf+prefixlen, TEXT("\r\n"), 2); // add trailing \r\n
 #endif
 	va_end(args);
